+++ apps/doc/src/app/components/navigation-menu/navigation-menu-example.module.ts
@@ -13,7 +13,7 @@ import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { NavigationMenuBasicExampleComponent } from './examples/navigation-menu-basic-example/navigation-menu-basic-example.component';
 import { NavigationMenuExampleComponent } from './navigation-menu-example.component';
-import { PrizmIconsSvgModule } from '@prizm-ui/icons';
+import { PrizmIconsSvgComponent } from '@prizm-ui/icons';
 import { NavigationMenuGroupsExampleComponent } from './examples/navigation-menu-groups-example/navigation-menu-groups-example.component';
 import { HintButtonComponent } from './examples/hint-button/hint-button.component';
 import { MenuHintComponent } from './examples/menu-hint/menu-hint.component';
@@ -34,7 +34,7 @@ import { NavigationMenuOneModeExampleComponent } from './examples/navigation-men
     CommonModule,
     PrizmAddonDocModule,
     PrizmNavigationMenuModule,
-    PrizmIconsSvgModule,
+    PrizmIconsSvgComponent,
     PrizmButtonModule,
     PrizmHintModule,
     PrizmBreadcrumbsModule,