+++ apps/doc/src/app/components/progress/line-segmented/examples/base/progress-base-example.component.ts
@@ -29,5 +29,5 @@ export class PrizmProgressBaseExampleComponent {
     takeWhile(i => i != this.max + 1),
     startWith(2)
   );
-  readonly colors = [`var(---prizm-v3-status-warning-primary-default)`, `lightskyblue`, `#3682db`, `red`];
+  readonly colors = [`var(---prizm-status-warning-primary-default)`, `lightskyblue`, `#3682db`, `red`];
 }