+++ apps/doc/src/app/logo/logo.module.ts
@@ -4,19 +4,28 @@ import { TuiLinkModule } from '@taiga-ui/core';
 import { LogoComponent } from './logo.component';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
-import { PrizmIconsSvgModule } from '@prizm-ui/icons';
-import { PrizmToggleModule } from '@prizm-ui/components';
+import { PrizmIconsSvgComponent } from '@prizm-ui/icons';
+import {
+  PrizmButtonComponent,
+  PrizmDialogComponent,
+  PrizmDialogModule,
+  PrizmToggleComponent,
+} from '@prizm-ui/components';
 import { PrizmLetDirective } from '@prizm-ui/helpers';
+import { ThemeTokenChangerComponent } from '../theme-token-changer/theme-token-changer.component';
 
 @NgModule({
   imports: [
     PrizmLetDirective,
     TuiLinkModule,
+    PrizmButtonComponent,
     CommonModule,
     RouterModule,
+    PrizmDialogModule,
+    ThemeTokenChangerComponent,
     FormsModule,
-    PrizmIconsSvgModule,
-    PrizmToggleModule,
+    PrizmIconsSvgComponent,
+    PrizmToggleComponent,
   ],
   declarations: [LogoComponent],
   exports: [LogoComponent],