+++ libs/components/src/lib/components/card/card.module.ts
@@ -1,7 +1,5 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { PrizmCardComponent } from './card.component';
-import { PrizmShadowModule } from '../../directives/shadow';
 
 /**
  * @deprecated