+++ libs/components/src/lib/components/input/input-date/input-layout-date.module.ts
@@ -2,10 +2,6 @@ import { NgModule } from '@angular/core';
 import { PrizmInputTextModule } from '../input-text/input-text.module';
 import { PrizmInputLayoutDateComponent } from './input-layout-date.component';
 
-/**
- * @deprecated
- * use standalone
- * */
 @NgModule({
   imports: [PrizmInputLayoutDateComponent],
   exports: [PrizmInputLayoutDateComponent, PrizmInputTextModule],