+++ libs/components/src/lib/components/input/input-month-range/input-layout-month-range.module.ts
@@ -2,10 +2,6 @@ import { NgModule } from '@angular/core';
 import { PrizmInputTextModule } from '../input-text/input-text.module';
 import { PrizmInputLayoutMonthRangeComponent } from './input-layout-month-range.component';
 
-/**
- * @deprecated
- * use standalone
- * */
 @NgModule({
   imports: [PrizmInputLayoutMonthRangeComponent, PrizmInputTextModule],
   exports: [PrizmInputLayoutMonthRangeComponent, PrizmInputTextModule],